# infrastructure/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Locação de Carros",
        "version": "0.1"
    },
    "paths": {
        "/api/carros/lista": {
            "get": {
                "summary": "Lista todos os carros",
                "description": "Retorna uma lista de todos os carros cadastrados",
                "operationId": "ea70398880412eb8972f0488d00cda7d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Carro"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/carros/cadastrar": {
            "post": {
                "summary": "Cadastrar um novo carro",
                "description": "Endpoint para cadastrar um novo carro",
                "operationId": "748f81fb37f37c9420c9c4773bf23add",
                "requestBody": {
                    "description": "Dados do carro a ser cadastrado",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "placa",
                                    "cor",
                                    "categoriaId"
                                ],
                                "properties": {
                                    "placa": {
                                        "type": "string",
                                        "example": "ABC1234"
                                    },
                                    "cor": {
                                        "type": "string",
                                        "example": "Preto"
                                    },
                                    "categoriaId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carro cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação dos dados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Parâmetros inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categorias/lista": {
            "get": {
                "summary": "Lista todas as categorias",
                "description": "Retorna uma lista de categorias de veículos",
                "operationId": "640ee2c431e1c82a42656b61220ab1e6",
                "responses": {
                    "200": {
                        "description": "Operação Concluída",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Categoria"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categorias/remover": {
            "delete": {
                "summary": "Remove uma categoria",
                "description": "Remove uma categoria pelo seu ID",
                "operationId": "62227e7589bb3b89a2d6b1631bcbf831",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID da categoria a ser removida",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação Concluída"
                    }
                }
            }
        },
        "/api/categorias/cadastrar": {
            "post": {
                "summary": "Cadastrar uma nova categoria",
                "description": "Endpoint para cadastrar um novo carro",
                "operationId": "b7a7bfa2057731d1c0aceab496074627",
                "requestBody": {
                    "description": "Dados do carro a ser cadastrado",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "taxa"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "ABC1234"
                                    },
                                    "taxa": {
                                        "type": "number",
                                        "format": "double",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carro cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação dos dados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Parâmetros inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/estacionamento/lista": {
            "get": {
                "summary": "Lista todos os carros estacionados",
                "description": "Retorna uma lista de todos os carros estacionados",
                "operationId": "0dc2fe2251fc951f778a7ded65c12134",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Estacionamento"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/estacionamento/enviar": {
            "post": {
                "summary": "Envia um entrada ou saída de um carro",
                "description": "Dá entrada em ou saída um carro no estacionamento",
                "operationId": "a7138fc84d6005ec0b85e0f7dcb3b1b1",
                "requestBody": {
                    "description": "Dados para a entrada",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "placa",
                                    "data",
                                    "park"
                                ],
                                "properties": {
                                    "placa": {
                                        "type": "string",
                                        "example": "000000"
                                    },
                                    "data": {
                                        "type": "string",
                                        "example": "2018-06-07T10:00"
                                    },
                                    "park": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entrada cadastrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação dos dados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Parâmetros inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Carro": {
                "title": "Carro",
                "description": "Entidade dos Carros",
                "properties": {
                    "placa": {
                        "description": "Placa do Veículo,",
                        "type": "string"
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/Categoria"
                    },
                    "cor": {
                        "description": "Cor do Veículo,",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Categoria": {
                "title": "Categoria",
                "description": "Entidade dos Categorias",
                "properties": {
                    "nome": {
                        "description": "Categoria de Veículo",
                        "type": "string"
                    },
                    "taxa": {
                        "description": "Taxa de Veículo",
                        "type": "string",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "Estacionamento": {
                "title": "Estacionamento",
                "description": "Entidade do Estacionamento",
                "properties": {
                    "entrada": {
                        "description": "Data de Entrada do Veículo,",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "saida": {
                        "description": "Data de Saída do Veículo,",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "park": {
                        "description": "Caso o Veículo Esteja Estacionado,",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    }
}